Switch back to native macOS application
This is heavily wip. Currently, only Spaces detection is implemented. Using a XPCService, IOSurfaces and the low-level XPCConnection implementation we are able to intercept Space captures and display them to the user (with a reasonable refresh rate, up to 30+fps [depending on the amount of Spaces].

Does not build currently, as compiling the XPCService emits a compiler error (which is most likely a compiler bug).

Signed-off-by: Philipp Remy <philipp.remy@law-school.de>