Fix unspecific Swift build error
When building the project with any prior commit, swift-frontend would fail mysterically with a non-zero exit code.

Turns out: The enums, which are passed to the XPCService to identify the request type cannot conform to BitwiseCopyable. Why? Idk.

Removing conformance for SpaceIOSurfaceCaptureServiceRequestType and SpaceIOSurfaceCaptureServiceResponseType to BitwiseCopyable fixes the compiler issue and causes the project to build correctly.

Signed-off-by: Philipp Remy <philipp.remy@law-school.de>